feat: add fluent validation

+++ Services/DimensionsService.cs
@@ -2,13 +2,13 @@ namespace CompanionApp.Services;
 
 public interface IDimensionsService
 {
-    double CalculateGap(double dimension, double marginLeft, double marginRight, int pointsCount);
+    double CalculateGap(int dimension, int marginLeft, int marginRight, int pointsCount);
 }
 
 public class DimensionsService : IDimensionsService
 {
-    public double CalculateGap(double dimension, double marginLeft, double marginRight, int pointsCount) =>
+    public double CalculateGap(int dimension, int marginLeft, int marginRight, int pointsCount) =>
         pointsCount != 1
-            ? (dimension - (marginLeft + marginRight)) / (pointsCount - 1)
-            : (dimension - (marginLeft + marginRight)) / 2; // sets the point right in the middle 
+            ? (double)(dimension - (marginLeft + marginRight)) / (pointsCount - 1)
+            : (double)(dimension - (marginLeft + marginRight)) / 2; // sets the point right in the middle 
 }
